Added rollup for build, added back docsify, cleaned up utils, created converter type and replaced usage
