+++ package.json
@@ -2,15 +2,18 @@
   "name": "n-color",
   "version": "1.0.0",
   "description": "A modern, fully fledged color toolkit",
-  "main": "dist/index.js",
+  "main": "dist/n-color.js",
+  "module": "dist/n-color.mjs",
+  "typings": "dist/n-color.d.ts",
   "scripts": {
     "start": "ts-node-dev --respawn --transpile-only src/example.ts",
     "build": "tsc -p tsconfig.prod.json",
+    "build:rollup": "rollup -c",
     "lint": "eslint \"src/**/*.ts\" \"src/**/*.test.ts\"",
     "test": "jest --config ./jest.config.ts",
     "test:watch": "jest --config ./jest.config.ts --watch",
     "docs:build": "documentation build dist/index.js -f html -o build/api",
-    "docs:serve": "docsify serve ./site",
+    "docs:serve": "docsify serve site",
     "docs:site": "mkdir -p build && rsync -a site/ build",
     "docs": "pnpm run docs:build && pnpm run docs:site",
     "docs:watch": "typedoc --watch"
@@ -27,6 +30,8 @@
     "@types/jest": "^27.4.0",
     "@typescript-eslint/eslint-plugin": "^5.9.1",
     "@typescript-eslint/parser": "^5.9.1",
+    "docsify-cli": "^4.4.4",
+    "esbuild": "^0.14.29",
     "eslint": "^8.2.0",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^16.1.0",
@@ -38,7 +43,11 @@
     "eslint-plugin-react": "^7.29.1",
     "eslint-plugin-react-hooks": "^4.3.0",
     "jest": "^27.4.7",
+    "markdownlint": "^0.25.1",
     "prettier": "^2.5.1",
+    "rollup": "^2.70.1",
+    "rollup-plugin-dts": "^4.2.0",
+    "rollup-plugin-esbuild": "^4.8.2",
     "ts-jest": "^27.1.2",
     "ts-node-dev": "^1.1.8",
     "typescript": "^4.5.4"
