+++ site/docs/_sidebar.md
@@ -1,3 +1,4 @@
 * [Home](/)
 * [Getting Started](getting-started.md)
+* [Getting Started](usage.md)
 * [API Documentation](/api)
