+++ src/spaces/hsl/index.ts
@@ -1,3 +1,2 @@
 export * as HSL from './hsl-resolver';
 export { default as HSLSpace } from './hsl-space';
-export { default as HSLColorSpace } from './types/hsl-color-space';
