+++ src/spaces/hsv/index.ts
@@ -1,3 +1,2 @@
-export * from './types';
 export { default as HSVSpace } from './hsv-space';
 export * as HSV from './hsv-resolver';
