--- src/spaces/hsv/types/index.ts
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import/prefer-default-export
-export { default as HSVColorSpace } from './hsv-space';
