+++ src/spaces/index.ts
@@ -1,5 +1,5 @@
-import { RGBA, RGBAColorSpace, RGBASpace } from './rgba';
-import { HSL, HSLColorSpace, HSLSpace } from './hsl';
+import { RGBA, RGBASpace } from './rgba';
+import { HSL, HSLSpace } from './hsl';
 
-export { RGBAColorSpace, RGBA, RGBASpace };
-export { HSLColorSpace, HSL, HSLSpace };
+export { RGBA, RGBASpace };
+export { HSL, HSLSpace };
