+++ src/spaces/rgba/index.ts
@@ -1,3 +1,2 @@
 export * as RGBA from './rgba-resolver';
 export { default as RGBASpace } from './rgba-space';
-export { default as RGBAColorSpace } from './types/rgba-color-space';
