+++ src/spaces/utils/index.ts
@@ -1,1 +1,3 @@
-export * from './color';
+export * from './hue-utils';
+export * from './hex-utils';
+export * from './converter';
